+++ packages/fireworks/src/fireworks-image-model.ts
@@ -69,6 +69,9 @@ interface FireworksImageModelConfig {
   baseURL: string;
   headers: () => Record<string, string>;
   fetch?: FetchFunction;
+  _internal?: {
+    currentDate?: () => Date;
+  };
 }
 
 const createBinaryResponseHandler =
@@ -179,7 +182,8 @@ export class FireworksImageModel implements ImageModelV1 {
     }
 
     const splitSize = size?.split('x');
-    const { value: response } = await postJsonToApi({
+    const currentDate = this.config._internal?.currentDate?.() ?? new Date();
+    const { value: response, responseHeaders } = await postJsonToApi({
       url: getUrlForModel(this.config.baseURL, this.modelId),
       headers: combineHeaders(this.config.headers(), headers),
       body: {
@@ -196,6 +200,14 @@ export class FireworksImageModel implements ImageModelV1 {
       fetch: this.config.fetch,
     });
 
-    return { images: [new Uint8Array(response)], warnings };
+    return {
+      images: [new Uint8Array(response)],
+      warnings,
+      response: {
+        timestamp: currentDate,
+        modelId: this.modelId,
+        headers: responseHeaders,
+      },
+    };
   }
 }