+++ packages/anthropic/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/anthropic
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies [ed012d2]
+- Updated dependencies [3a58a2e]
+  - @ai-sdk/provider-utils@2.1.2
+  - @ai-sdk/provider@1.0.6
+
+## 1.1.1
+
+### Patch Changes
+
+- 858f934: feat (provider/anthropic): default cache-control on and mark model setting deprecated
+- b284e2c: feat (provider/google-vertex): support prompt caching for Anthropic Claude models
+- Updated dependencies [e7a9ec9]
+- Updated dependencies [0a699f1]
+  - @ai-sdk/provider-utils@2.1.1
+  - @ai-sdk/provider@1.0.5
+
 ## 1.1.0
 
 ### Minor Changes