+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/deepseek
 
+## 0.1.2
+
+### Patch Changes
+
+- ed012d2: feat (provider/deepseek): extract cache usage as provide metadata
+- Updated dependencies [ed012d2]
+- Updated dependencies [3a58a2e]
+  - @ai-sdk/openai-compatible@0.1.2
+  - @ai-sdk/provider-utils@2.1.2
+  - @ai-sdk/provider@1.0.6
+
+## 0.1.1
+
+### Patch Changes
+
+- 0a699f1: feat: add reasoning token support
+- Updated dependencies [e7a9ec9]
+- Updated dependencies [0a699f1]
+  - @ai-sdk/provider-utils@2.1.1
+  - @ai-sdk/openai-compatible@0.1.1
+  - @ai-sdk/provider@1.0.5
+
 ## 0.1.0
 
 ### Minor Changes