+++ packages/swarm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/swarm
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [3c5fafa]
+- Updated dependencies [3a58a2e]
+  - ai@4.1.2
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [0a699f1]
+  - ai@4.1.1
+
 ## 0.1.0
 
 ### Minor Changes